pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "surrogate-ccm"
version = "0.1.0"
description = "Surrogate-Enhanced CCM for Causal Inference in Complex Networks"
requires-python = ">=3.9"
dependencies = [
    "numpy>=1.24",
    "scipy>=1.10",
    "scikit-learn>=1.2",
    "networkx>=3.0",
    "matplotlib>=3.7",
    "seaborn>=0.12",
    "joblib>=1.3",
    "h5py>=3.8",
    "pyyaml>=6.0",
    "tqdm>=4.65",
    "statsmodels>=0.14",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-cov"]

[tool.setuptools.packages.find]
include = ["surrogate_ccm*"]
